ai-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
vanna[postgres]==0.5.5
groq==0.4.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pandas==2.1.4
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1
jinja2==3.1.2